Finished user-system: users stay logged index for 10 min, can only edit and delete their own articles only users can create articles, still need to implement authentication for creating an article

+++ server/src/server.js
@@ -33,7 +33,7 @@ kategoridao = new KategoriDao(pool);
 kommentardao = new KommentarDao(pool);
 ratingdao = new RatingDao(pool);
 
-app.use(function(req, res, next) {
+app.use(function (req, res, next) {
     res.header("Access-Control-Allow-Origin", "http://localhost:3000"); // update to match the domain you will make the request from
     res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
     next();
@@ -60,12 +60,12 @@ app.get("/api/nyheter", (req, res) => {
     sakdao.getNyheter((status, data) => {
         res.status(status);
         res.json(data);
-      });
+    });
 });
 
 app.get("/api/nyheter/:sak_id", (req, res) => {
     console.log("Fikk request fra klient");
-    sakdao.getNyhet(req.params.sak_id, (status, data)=>{
+    sakdao.getNyhet(req.params.sak_id, (status, data) => {
         res.status(status);
         res.json(data);
     });
@@ -103,33 +103,70 @@ app.get("/api/kategori/:kategori",(req, res) => {
 app.post("/api/nyheter", (req, res) => {
     console.log("Fikk POST-request fra klienten");
     console.log("Navn: " + req.body.navn);
-    sakdao.lagNyhet(req.body, (status, data)=>{
+    sakdao.lagNyhet(req.body, (status, data) => {
         res.status(status);
         res.json(data);
     });
 });
 
 app.delete("/api/nyheter/", (req, res) => {
-    console.log("Fikk DELETE-request fra klienten");
-    sakdao.deleteNyhet(req.body, (status, data)=>{
-        res.status(status);
-        res.json(data);
-    });
-  
+    //console.log("Fikk DELETE-request fra klienten");
+    var token = req.headers["x-access-token"];
+    jwt.verify(token, privateKEY.key, (err, decoded)=>{
+        if(err){
+            res.status(401);
+            res.json({error: "Not authorized"});
+        }else{
+            sakdao.getBrukernavnAvSak_id(req.body.sak_id, (status, data)=>{
+                if(data[0].brukernavn == decoded.brukernavn){
+                    sakdao.deleteNyhet(req.body, (status, data)=>{
+                        res.status(status);
+                        let token = jwt.sign({ brukernavn: decoded.brukernavn }, privateKEY.key, {
+                            expiresIn: 6000
+                        });
+                        res.json({jwt:token});
+                    })
+                }else{
+                    res.status(401);
+                    res.json({ error: "Not authorized"});    
+                }
+            })
+        }
+    })
 });
 
 app.put("/api/nyheter/:sak_id", (req, res) => {
-    console.log("Fikk PUT-request fra klienten");
-    console.log("Navn: " + req.body.navn);
-    sakdao.oppdaterNyhet(req.body, req.params.sak_id, (status, data)=>{
-        res.status(status);
-        res.json(data);
-    });
+    var token = req.headers["x-access-token"];
+    console.log("token: "+ token);
+    jwt.verify(token, privateKEY.key, (err, decoded) => {
+        if (err) {
+            console.log("token ikke ok");
+            res.status(401);
+            res.json({ error: "Not authorized" });
+        } else {
+            console.log("Token ok: " + decoded.brukernavn);
+            sakdao.getBrukernavnAvSak_id(req.params.sak_id, (status, data)=>{
+                console.log(data[0].brukernavn);
+                if(data[0].brukernavn == decoded.brukernavn){
+                    sakdao.oppdaterNyhet(req.body, req.params.sak_id,(status, data)=>{
+                        res.status(status);
+                        let token = jwt.sign({ brukernavn: decoded.brukernavn }, privateKEY.key, {
+                            expiresIn: 6000
+                        });
+                        res.json({ jwt: token });
+                    })
+                }else{
+                    res.status(401);
+                    res.json({ error: "Not authorized"});
+                }
+            })
+        }
+    })
 });
 
 app.get("/api/kategorier", (req, res) => {
     console.log("Fikk request fra klient");
-    kategoridao.getKategorier((status, data)=>{
+    kategoridao.getKategorier((status, data) => {
         res.status(status);
         res.json(data);
     });
@@ -137,7 +174,7 @@ app.get("/api/kategorier", (req, res) => {
 
 app.get("/api/kommentar/:sak_id", (req, res) => {
     console.log("Fikk request fra klient");
-    kommentardao.getKommentar(req.params.sak_id, (status, data)=>{
+    kommentardao.getKommentar(req.params.sak_id, (status, data) => {
         res.status(status);
         res.json(data);
     })
@@ -146,7 +183,7 @@ app.get("/api/kommentar/:sak_id", (req, res) => {
 app.post("/api/kommentar", (req, res) => {
     console.log("Fikk POST-request fra klienten");
     console.log("Navn: " + req.body.navn);
-    kommentardao.addKommentar(req.body, (status, data)=>{
+    kommentardao.addKommentar(req.body, (status, data) => {
         res.status(status);
         res.json(data);
     });
@@ -155,7 +192,7 @@ app.post("/api/kommentar", (req, res) => {
 app.post("/api/rating/:sak_id", (req, res) => {
     console.log("Fikk POST-request fra klienten");
     console.log("sakid: " + req.params.sak_id);
-    ratingdao.addRating(req.body, req.params.sak_id, (status, data)=>{
+    ratingdao.addRating(req.body, req.params.sak_id, (status, data) => {
         res.status(status);
         res.json(data);
     });
@@ -163,7 +200,7 @@ app.post("/api/rating/:sak_id", (req, res) => {
 app.get("/api/rating/:sak_id", (req, res) => {
     console.log("Fikk GET-request fra klienten");
     console.log("sakid: " + req.params.sak_id);
-    ratingdao.getRating(req.params.sak_id, (status, data)=>{
+    ratingdao.getRating(req.params.sak_id, (status, data) => {
         res.status(status);
         res.json(data);
     });
@@ -172,37 +209,57 @@ app.get("/api/rating/:sak_id", (req, res) => {
 app.post("/api/bruker", (req, res) => {
     console.log("Fikk POST-request fra klienten");
     bdao.addBruker(req.body, (status, data) => {
-      res.status(status);
-      res.json(data);
-    });
-  });
-
-app.post("/api/login",(req,res)=>{
-    console.log("Fikk POST-request fra klienten");
-    bdao.getBruker(req.body, (status, data)=>{
         res.status(status);
-        if(data[0]){
-            bcrypt.compare(req.body.passord, data[0].passord, function(err, resp) {
-                if(resp) {
-                    let token = jwt.sign({brukernavn: req.body.brukernavn}, privateKEY.key,{
+        res.json(data);
+    });
+});
+
+app.post("/api/login", (req, res) => {
+    console.log("Fikk POST-request fra klienten");
+    bdao.getBruker(req.body, (status, data) => {
+        res.status(status);
+        if (data[0]) {
+            bcrypt.compare(req.body.passord, data[0].passord, function (err, resp) {
+                if (resp) {
+                    let token = jwt.sign({ brukernavn: req.body.brukernavn }, privateKEY.key, {
                         expiresIn: 600
                     });
                     console.log("password matched");
                     res.status(status);
-                    res.json({jwt: token});
+                    res.json({ jwt: token });
                 } else {
                     console.log("password didnt match");
                     res.status(401);
-                    res.json({error: "not authorized" });
-                } 
-              });
-        }else{
+                    res.json({ error: "not authorized" });
+                }
+            });
+        } else {
             res.status(401);
-            res.json({error: "user does not exist"});
+            res.json({ error: "user does not exist" });
         }
-        });
+    });
 });
+
+app.post("/token", (req, res)=>{
+    var token = req.headers["x-access-token"];
+    console.log("token: "+ token )
+    jwt.verify(token, privateKEY.key, (err, decoded)=>{
+      if(err){
+        console.log("Token utløpt");
+        res.status(401);
+        res.json({error: "not authorized"});
+      }else{
+        console.log("Token ok: " + decoded.brukernavn);
+        console.log("token refreshed");
+        token = jwt.sign({brukernavn: decoded.brukernavn}, privateKEY.key, {
+          expiresIn:6000
+        });
+        res.json({jwt: token, "brukernavn": decoded.brukernavn});
+      }
+    })
+  })
   
+
